+++ apm-sniffer/apm-sdk-plugin/jersey-1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/jersey/v1/Test.java
@@ -1,5 +1,0 @@
-package org.apache.skywalking.apm.plugin.jersey.v1;
-
-public class Test {
-
-}