+++ CHANGES.md
@@ -8,6 +8,7 @@ Release Notes.
 * Exclude `synthetic` methods for the WitnessMethod mechanism
 * Support ForkJoinPool trace
 * Support clickhouse-jdbc-plugin trace sql parameters
+* Support Jersey REST framework
 
 #### Documentation